+++ packages/tsurlfilter/src/rules/declarative-converter/grouped-rules-converters/abstract-rule-converter.ts
@@ -609,6 +609,7 @@ export abstract class DeclarativeRuleConverter {
          * also for all other types of requests.
          */
         const shouldMatchAllResourcesTypes = rule.isOptionEnabled(NetworkRuleOption.RemoveHeader)
+            || rule.isOptionEnabled(NetworkRuleOption.RemoveParam)
             || rule.isOptionEnabled(NetworkRuleOption.Csp)
             || rule.isOptionEnabled(NetworkRuleOption.Permissions)
             || rule.isOptionEnabled(NetworkRuleOption.Cookie)